Bug Fixes (#40)
* updating documentation link that was changed because there is no longer a config.yaml

* wrapping all of excel_uploader_main.py in a try catch from gui_main.py

* wrapping all of excel_uploader_main.py in a try catch from gui_main.py

* clarified comment in the section of try catch of len(error_list)

* added except Exception as error

* added small clarification as to why user is going to globus screen

* added comment as to why setuptools is imported but not used for better code readability

* changed version from 0.6.0 to 0.6.1

* added an upload again button to the success screen as well as error screen

* underlined globus auth button link to indicate to the user that the globus auth button is a link

* Exceptions no longer exist because of SDK refactoring, so they have been changed to generic for now and it is working

* turning cursor into "not-allowed" icon for Excel file path input

* fixed start_screen having broken feedback

all_screens.js: put restartStartScreen.js into a function that can be used by restartAllScreens function and start_screen.js submitForm() function

start_screen.js: clearing the start_screen form validations before sending user input to not get any leftover/wrong feedback after submission
