+++ src/web/templates/base.html
@@ -43,7 +43,7 @@
     <!-- my styles -->
     <link href="../css/styles.css" rel="stylesheet">
 
-    <title>CRIPT Excel Uploader - version 0.6.0</title>
+    <title>CRIPT Excel Uploader - version 0.6.1</title>
 </head>
 <body>
 <header>
@@ -155,7 +155,8 @@
                     Excel File:
                 </button>
 
-                <input type="text" id="excel-file-path" class="form-control"
+                <input type="text" id="excel-file-path"
+                       class="form-control" style="cursor: not-allowed;"
                        placeholder="Excel absolute file path"
                        aria-label="place host here"
                        aria-describedby="excel-file-button excel-file-path-validation-feedback"
@@ -181,19 +182,19 @@
                     Upload
                 </button>
             </div>
-
-            <br>
-
-            <!-- docs link for how to fill out config -->
-            <small class="documentation-link">
-                Please refer to our
-
-                <a href="https://c-accel-cript.github.io/cript-excel-uploader/filling_out_config/" target="_blank">
-                    documentation
-                </a>
-                on how to fill out this screen
-            </small>
         </form>
+
+        <br>
+
+        <!-- docs link for how to fill out config -->
+        <small class="documentation-link">
+            Please refer to our
+
+            <a href="https://c-accel-cript.github.io/cript-excel-uploader/filling_out_first_screen/" target="_blank">
+                documentation
+            </a>
+            on how to fill out this screen
+        </small>
     </div>
 
     <!-- LOADING SCREEN -->
@@ -258,7 +259,7 @@
 
         <!-- screen header that explains the screen -->
         <h2 style="font-size: 1.5rem;" class="text-center">
-            Globus Authentication for File Uploads
+            Globus Authentication for <u>Local</u> File Uploads
         </h2>
 
         <!-- image -->
@@ -277,7 +278,8 @@
 
         <!-- Globus Auth Link -->
         <div class="text-center" style="margin-top: 1rem;">
-            <a id="globus-auth-link" class="btn btn-primary " href="" target="_blank">
+            <a id="globus-auth-link" class="btn btn-primary" style="text-decoration: underline"
+               href="" target="_blank">
                 <i class="material-icons">vpn_lock</i>
                 Globus Authentication
             </a>
@@ -344,7 +346,7 @@
 
         <!-- Try again button -->
         <div class="d-grid gap-2" style="margin-top: 2rem;">
-            <button type="button" onclick="tryAgain();"
+            <button type="button" onclick="uploadAgain();"
                     class="btn btn-primary btn-lg margin-top d-flex justify-content-center"
                     style="padding: 0.7rem 7rem 0.7rem 7rem;">
                 Try again
@@ -378,12 +380,18 @@
         <img src="../assets/images/done.png" class="img-fluid" style="height: 30rem;"
              alt="cartoon picture of task completed">
 
+        <!-- success and upload again button -->
         <div class="d-grid gap-2">
-            <!-- onclick close the window -->
-            <a id="data-in-cript-link" class="btn" href="https://criptapp.org" target="_blank">
+            <!-- show uploaded collection -->
+            <a id="data-in-cript-link" class="btn btn-link" href="https://criptapp.org" target="_blank">
                 <i class="material-icons">cloud_done</i>
                 See my data in CRIPT
             </a>
+            <!-- upload again -->
+            <button class="btn btn-primary" type="button" onclick="uploadAgain()">
+                <i class="material-icons">restart_alt</i>
+                Upload again
+            </button>
         </div>
     </div>
